% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top10_by_country_and_year.R
\name{plot_top10_by_country_and_year}
\alias{plot_top10_by_country_and_year}
\title{Create a plot with top10 country of orgin for Refugees and OIP per country of asylum and year}
\usage{
plot_top10_by_country_and_year(year, countryname)
}
\arguments{
\item{year}{a numeric value for the year}

\item{countryname}{the name of the country of asulum}
}
\value{
a ggplot2 with UNHCR brand
}
\description{
Create a plot with top10 country of orgin for Refugees and OIP per country of asylum and year
}
\examples{
plot_top10_by_country_and_year( year = 2022, 
                                 countryname = "Peru")
}
